radeon_replay: Move linux-centric helpers to separate files
For ease of porting, we use data structures that mimic the ones from
linux. To allow those structures to be used in more than one file,
move them out of aruba_dp_aux.c, and into their own files.
